--- src/components/organism/HeaderSection.tsx
@@ -1,43 +1,0 @@
-import styled, { css } from 'styled-components';
-import BaseLayoutProps from '../types/BaseLayoutProps';
-import * as React from 'react';
-
-export interface HeaderProps extends BaseLayoutProps {}
-
-const HeaderStyled = styled.header`
-  ${({ theme }) => css`
-    background-color: ${theme.colors.purple};
-  `}
-`;
-
-export default function HeaderSection() {
-  return (
-    <HeaderStyled>
-      <svg
-        width="242"
-        height="40"
-        viewBox="0 0 242 40"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          d="M25.0872 16H30.9128C31.5133 16 32 15.5133 32 14.9128V9.08716C32 8.48675 31.5133 8 30.9128 8H25.0872C24.4867 8 24 8.48675 24 9.08716V14.9128C24 15.5133 24.4867 16 25.0872 16Z"
-          fill="white"
-        />
-        <path
-          d="M17.0872 24H22.9128C23.5133 24 24 23.5133 24 22.9129V17.0872C24 16.4867 23.5133 16 22.9128 16H17.0872C16.4867 16 16 16.4867 16 17.0872V22.9129C16 23.5133 16.4867 24 17.0872 24Z"
-          fill="white"
-        />
-        <path
-          d="M9.08715 32H14.9129C15.5133 32 16 31.5133 16 30.9128V25.0871C16 24.4867 15.5133 24 14.9129 24H9.08715C8.48673 24 8 24.4867 8 25.0871V30.9128C8 31.5133 8.48673 32 9.08715 32Z"
-          fill="white"
-        />
-        <path
-          d="M46.0753 14.2707V27H48.636V18.634H48.7417L52.0608 26.9441H53.826L57.145 18.6651H57.2507V27H59.8177V14.2707H56.5546L53.018 22.8978H52.8688L49.3322 14.2707H46.0753ZM61.4562 24.3336C61.4562 26.2169 62.8112 27.1802 64.6323 27.1802C66.0495 27.1802 66.9631 26.5587 67.4231 25.6761H67.4977V27H69.9652V20.567C69.9652 18.2673 68.0011 17.3287 65.9189 17.3287C63.383 17.3287 61.941 18.634 61.7919 20.2935H64.2594C64.4024 19.6409 64.9618 19.2369 65.8568 19.2369C66.8202 19.2369 67.3671 19.6968 67.3734 20.5048V21.2756C66.9756 21.288 65.6828 21.3439 65.0053 21.3874C63.1717 21.5055 61.4562 22.1954 61.4562 24.3336ZM63.9486 24.2528C63.9486 23.5193 64.5205 23.0905 65.4279 23.0159C65.8754 22.9724 67.0004 22.9102 67.3858 22.8916V23.5877C67.392 24.5946 66.5591 25.3715 65.372 25.3715C64.5391 25.3715 63.9486 24.9924 63.9486 24.2528ZM71.9409 27H74.5451V21.5739C74.5451 20.3992 75.4091 19.5787 76.5838 19.5787C76.9505 19.5787 77.454 19.6471 77.6964 19.7217V17.3971C77.4664 17.3536 77.1308 17.3163 76.8635 17.3163C75.7944 17.3163 74.9119 17.9378 74.5638 19.1126H74.4643V17.453H71.9409V27ZM78.9721 27H81.5764V23.9482L82.3036 23.1215L84.9452 27H87.997L84.249 21.5863L87.8105 17.453H84.8209L81.7193 21.1015H81.5764V14.2707H78.9721V27ZM92.9371 27.1865C95.2928 27.1865 96.8715 26.0428 97.2444 24.2838H94.7707C94.5096 24.8494 93.8756 25.2348 92.9806 25.2348C91.6443 25.2348 90.7865 24.346 90.7865 22.8978H97.3128V22.1706C97.3128 18.9137 95.3363 17.3287 92.8314 17.3287C90.0345 17.3287 88.2133 19.3239 88.2133 22.27C88.2133 25.297 90.0096 27.1865 92.9371 27.1865ZM90.7865 21.2693C90.8487 20.1754 91.6816 19.2742 92.8812 19.2742C94.0435 19.2742 94.8453 20.1133 94.8577 21.2693H90.7865ZM103.982 17.453H102.179V15.1657H99.5748V17.453H98.2634V19.4109H99.5748V24.4268C99.5624 26.2852 100.843 27.1305 102.732 27.1305C103.366 27.1305 103.832 27.0808 104.106 27.0373V24.9862C103.795 25.0048 103.341 25.0359 103.087 25.0297C102.552 25.011 102.179 24.8308 102.179 24.0725V19.4109H103.982V17.453ZM109.814 14.2707V27H112.374V18.634H112.48L115.799 26.9441H117.564L120.883 18.6651H120.989V27H123.556V14.2707H120.293L116.756 22.8978H116.607L113.07 14.2707H109.814ZM129.943 27.1865C132.827 27.1865 134.623 25.2037 134.623 22.2638C134.623 19.3052 132.827 17.3287 129.943 17.3287C127.053 17.3287 125.25 19.3052 125.25 22.2638C125.25 25.2037 127.053 27.1865 129.943 27.1865ZM127.898 22.2452C127.898 20.5732 128.594 19.3177 129.956 19.3177C131.286 19.3177 131.982 20.5732 131.982 22.2452C131.982 23.9233 131.286 25.1664 129.956 25.1664C128.594 25.1664 127.898 23.9233 127.898 22.2452ZM138.863 21.4744C138.863 20.2189 139.609 19.4917 140.697 19.4917C141.772 19.4917 142.413 20.2003 142.406 21.375V27H145.011V20.9213C145.011 18.6899 143.712 17.3287 141.723 17.3287C140.299 17.3287 139.28 18.0249 138.851 19.1312H138.739V17.453H136.259V27H138.863V21.4744ZM147.017 27H149.621V17.453H147.017V27ZM146.905 14.8798C146.905 15.6133 147.545 16.2037 148.322 16.2037C149.099 16.2037 149.739 15.6133 149.739 14.8798C149.739 14.1526 149.099 13.5559 148.322 13.5559C147.545 13.5559 146.905 14.1526 146.905 14.8798ZM156.681 17.453H154.878V15.1657H152.274V17.453H150.963V19.4109H152.274V24.4268C152.262 26.2852 153.542 27.1305 155.432 27.1305C156.065 27.1305 156.532 27.0808 156.805 27.0373V24.9862C156.494 25.0048 156.041 25.0359 155.786 25.0297C155.251 25.011 154.878 24.8308 154.878 24.0725V19.4109H156.681V17.453ZM162.529 27.1865C164.885 27.1865 166.463 26.0428 166.836 24.2838H164.362C164.101 24.8494 163.467 25.2348 162.572 25.2348C161.236 25.2348 160.378 24.346 160.378 22.8978H166.905V22.1706C166.905 18.9137 164.928 17.3287 162.423 17.3287C159.626 17.3287 157.805 19.3239 157.805 22.27C157.805 25.297 159.601 27.1865 162.529 27.1865ZM160.378 21.2693C160.44 20.1754 161.273 19.2742 162.473 19.2742C163.635 19.2742 164.437 20.1133 164.449 21.2693H160.378ZM168.533 27H171.137V21.5739C171.137 20.3992 172.001 19.5787 173.176 19.5787C173.542 19.5787 174.046 19.6471 174.288 19.7217V17.3971C174.058 17.3536 173.723 17.3163 173.455 17.3163C172.386 17.3163 171.504 17.9378 171.156 19.1126H171.056V17.453H168.533V27ZM175.564 27H178.168V17.453H175.564V27ZM175.452 14.8798C175.452 15.6133 176.092 16.2037 176.869 16.2037C177.646 16.2037 178.286 15.6133 178.286 14.8798C178.286 14.1526 177.646 13.5559 176.869 13.5559C176.092 13.5559 175.452 14.1526 175.452 14.8798ZM182.791 21.4744C182.791 20.2189 183.537 19.4917 184.625 19.4917C185.7 19.4917 186.34 20.2003 186.334 21.375V27H188.938V20.9213C188.938 18.6899 187.639 17.3287 185.65 17.3287C184.227 17.3287 183.208 18.0249 182.779 19.1312H182.667V17.453H180.187V27H182.791V21.4744ZM195.233 30.779C197.95 30.779 199.87 29.5421 199.87 27.0994V17.453H197.297V19.0566H197.191C196.837 18.2859 196.073 17.3287 194.463 17.3287C192.356 17.3287 190.584 18.9385 190.584 22.1706C190.584 25.3467 192.312 26.8633 194.469 26.8633C196.004 26.8633 196.843 26.0863 197.198 25.3032H197.309V27.0622C197.309 28.386 196.501 28.8957 195.289 28.8957C194.258 28.8957 193.537 28.5539 193.394 27.7831H190.827C190.982 29.5981 192.604 30.779 195.233 30.779ZM193.244 22.1954C193.244 20.5546 193.947 19.4109 195.283 19.4109C196.582 19.4109 197.316 20.5048 197.316 22.1954C197.316 23.8923 196.57 24.9054 195.283 24.9054C193.959 24.9054 193.244 23.855 193.244 22.1954ZM205.603 16.4586H209.513V27H212.129V16.4586H216.026V14.2707H205.603V16.4586ZM220.523 27.1865C223.407 27.1865 225.203 25.2037 225.203 22.2638C225.203 19.3052 223.407 17.3287 220.523 17.3287C217.633 17.3287 215.831 19.3052 215.831 22.2638C215.831 25.2037 217.633 27.1865 220.523 27.1865ZM218.478 22.2452C218.478 20.5732 219.174 19.3177 220.536 19.3177C221.866 19.3177 222.562 20.5732 222.562 22.2452C222.562 23.9233 221.866 25.1664 220.536 25.1664C219.174 25.1664 218.478 23.9233 218.478 22.2452ZM231.14 27.1865C234.024 27.1865 235.821 25.2037 235.821 22.2638C235.821 19.3052 234.024 17.3287 231.14 17.3287C228.25 17.3287 226.448 19.3052 226.448 22.2638C226.448 25.2037 228.25 27.1865 231.14 27.1865ZM229.096 22.2452C229.096 20.5732 229.792 19.3177 231.153 19.3177C232.483 19.3177 233.179 20.5732 233.179 22.2452C233.179 23.9233 232.483 25.1664 231.153 25.1664C229.792 25.1664 229.096 23.9233 229.096 22.2452ZM240.061 14.2707H237.456V27H240.061V14.2707Z"
-          fill="white"
-        />
-      </svg>
-    </HeaderStyled>
-  );
-}
-// TODO: 추후 TextBtn 추가
